Move our embedded postgres test into a proper test suite, and indicate that it is an integration test
