+++ model/model_test.go
@@ -1,18 +1,10 @@
 package model_test
 
 import (
-	"log"
-	"path"
-	"path/filepath"
-	"runtime"
 	"testing"
 
 	"github.com/calvinchengx/gin-go-pg/mock"
 	"github.com/calvinchengx/gin-go-pg/model"
-	embeddedpostgres "github.com/fergusstrange/embedded-postgres"
-	"github.com/go-pg/pg/v9"
-	"github.com/go-pg/pg/v9/orm"
-	"github.com/stretchr/testify/assert"
 )
 
 func TestBeforeInsert(t *testing.T) {
@@ -51,44 +43,4 @@ func TestDelete(t *testing.T) {
 		t.Errorf("DeletedAt not changed")
 	}
 
-}
-
-func TestDatabase(t *testing.T) {
-	_, b, _, _ := runtime.Caller(0)
-	d := path.Join(path.Dir(b))
-	projectRoot := filepath.Dir(d)
-	tmpDir := path.Join(projectRoot, "tmp")
-	testConfig := embeddedpostgres.DefaultConfig().
-		Username("db_test_user").
-		Password("db_test_password").
-		Database("db_test_database").
-		Version("12.1.0").
-		RuntimePath(tmpDir).
-		Port(9876)
-
-	postgres := embeddedpostgres.NewDatabase(testConfig)
-	err := postgres.Start()
-	assert.Equal(t, err, nil)
-
-	db := pg.Connect(&pg.Options{
-		Addr:     "localhost:9876",
-		User:     "db_test_user",
-		Password: "db_test_password",
-		Database: "db_test_database",
-	})
-	createSchema(db, &model.Company{}, &model.Location{}, &model.Role{}, &model.User{}, &model.Verification{})
-
-	_ = postgres.Stop()
-}
-
-func createSchema(db *pg.DB, models ...interface{}) {
-	for _, model := range models {
-		opt := &orm.CreateTableOptions{
-			IfNotExists: true,
-		}
-		err := db.CreateTable(model, opt)
-		if err != nil {
-			log.Fatal(err)
-		}
-	}
 }
